requirements.txt
certifi
chardet
click
colorama
colorzero
Flask
gpiozero
idna
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
numpy
pandas
Pillow
python-dateutil
pytz
requests
six
typing-extensions
urllib3
Werkzeug
zipp
twilio
opencv-python
pygame